Widen pdf granularity to int64 and cap forward-algorithm step count
Two fixes for graph.pdf() / graph.cdf() on phase-type models with extreme
edge rates:

1. granularity is now int64_t throughout the C struct, C function signature,
   C++ wrapper methods, and pybind binding (was int32, which silently rejected
   values like 1.5e11 with a cryptic TypeError).

2. The C++ wrapper now caps the forward-algorithm step count at 1e9 (~16 GB
   of _pdf/_cdf history) before entering the discretization loop. The check
   uses ph_context->granularity * time, so it covers both explicit-granularity
   calls and the granularity=0 auto-selection path. Without the cap, models
   with large rate spreads (e.g. uranium decay: 21 orders of magnitude
   between fastest and slowest transition) would consume tens of GB of
   memory before the user could intervene.

Failure mode now raises std::invalid_argument with concrete numbers
(estimated step count, resolved granularity, time) pointing the user at
the cause.

+++ api/c/phasic.h
@@ -994,7 +994,7 @@ struct ptd_probability_distribution_context {
     struct ptd_graph *graph;
     void *priv;
     long double time;
-    int granularity;
+    int64_t granularity;
     /* graph->weight_version snapshot taken when this context was created.
      * Used by C++ wrapper cache guards (api/cpp/phasiccpp.h) to detect
      * stale forward state after edge weights change. */
@@ -1003,7 +1003,7 @@ struct ptd_probability_distribution_context {
 
 struct ptd_probability_distribution_context *ptd_probability_distribution_context_create(
         struct ptd_graph *graph,
-        int granularity
+        int64_t granularity
 );
 
 void ptd_probability_distribution_context_destroy(
